stremio_core::models
Module meta_details

Structs

MetaDetails

Selected

Functions

last_used_stream_update 🔒

Find a stream from addon responses, which should be played if binge watching or continuing to watch. We’ve already loaded the next Video id and we need to find a proper stream for the binge watching.

library_item_sync 🔒

library_item_update 🔒

meta_items_update 🔒

meta_streams_update 🔒

selected_guess_stream_update 🔒

If Selected::guess_stream is true then we will override the selected stream no matter if it’s set (Some) or not (None).

streams_update 🔒

watched_update 🔒