Struct stremio_core::runtime::effects::Effects

pub struct Effects {
    effects: Vec<Effect>,
    pub has_changed: bool,
}

Examples

use stremio_core::runtime::Effects;

let none_unchanged = Effects::none().unchanged();
let default = Effects::default();

assert_eq!(default.has_changed, false);
assert!(none_unchanged.is_empty());
assert!(default.is_empty());
assert_eq!(none_unchanged.has_changed, default.has_changed);

Fields
effects: Vec<Effect>

The effects to be applied

has_changed: bool

Whether or not the effects are changing something in the Model.

When has_changed == true it will mark the given model field as changed and add effect for it. This will trigger a RuntimeEvent::NewState for the Model with all it’s changed fields.

Implementations

impl Effects

pub fn none() -> Self

pub fn one(effect: Effect) -> Self

pub fn many(effects: Vec<Effect>) -> Self

pub fn msg(msg: Msg) -> Self

pub fn future(future: EffectFuture) -> Self

pub fn msgs(msgs: Vec<Msg>) -> Self

pub fn futures(futures: Vec<EffectFuture>) -> Self

pub fn unchanged(self) -> Self

mark model as not changed

pub fn join(self, effects: Effects) -> Self

pub fn len(&self) -> usize

Returns the amount of effects

pub fn is_empty(&self) -> bool

Returns whether there are any effects