Module stremio_core::runtime::msg

Modules

action 🔒

event 🔒

internal 🔒

msg 🔒

Structs

CtxAuthResponse

PlayOnDeviceArgs

Enums

Action

Action messages

ActionCatalogWithFilters

ActionCatalogsWithExtra

ActionCtx

ActionLibraryByType

ActionLibraryWithFilters

ActionLink

ActionLoad

ActionMetaDetails

ActionPlayer

ActionSearch

ActionStreamingServer

CreateTorrentArgs

Event

Those messages are meant to be dispatched by the stremio-core crate and handled by the users of the stremio-core crate and by the stremio-core crate itself.

Internal

Msg

Type Aliases

CtxStorageResponse

LibraryPlanResponse