Enum stremio_core::runtime::msg::action::ActionCtx

pub enum ActionCtx {

    Authenticate(AuthRequest),
    Logout,
    InstallAddon(Descriptor),
    InstallTraktAddon,
    LogoutTrakt,
    UpgradeAddon(Descriptor),
    UninstallAddon(Descriptor),
    UpdateSettings(Settings),
    AddToLibrary(MetaItemPreview),
    RemoveFromLibrary(String),
    RewindLibraryItem(String),
    LibraryItemMarkAsWatched {
        id: LibraryItemId,
        is_watched: bool,
    },
    ToggleLibraryItemNotifications(LibraryItemId, bool),
    DismissNotificationItem(MetaItemId),
    ClearSearchHistory,
    PushUserToAPI,
    PullUserFromAPI,
    PushAddonsToAPI,
    PullAddonsFromAPI,
    SyncLibraryWithAPI,
    PullNotifications,
    GetEvents,
    DismissEvent(String),

}

Variants

Authenticate(AuthRequest)

Logout

InstallAddon(Descriptor)

InstallTraktAddon

LogoutTrakt

UpgradeAddon(Descriptor)

UninstallAddon(Descriptor)

UpdateSettings(Settings)

AddToLibrary(MetaItemPreview)

RemoveFromLibrary(String)

RewindLibraryItem(String)

LibraryItemMarkAsWatched

Fields

id: LibraryItemId

is_watched: bool

ToggleLibraryItemNotifications(LibraryItemId, bool)

If boolean is set to true it will disable notifications for the LibraryItem.

DismissNotificationItem(MetaItemId)

Dismiss all Notification for a given MetaItemId.

ClearSearchHistory

PushUserToAPI

PullUserFromAPI

PushAddonsToAPI

PullAddonsFromAPI

SyncLibraryWithAPI

PullNotifications

Pull notifications for all LibraryItems that we should pull notifications for.

Warning: The action will always trigger requests to the addons.

See LibraryItem::should_pull_notifications()

GetEvents

Make request to api to get events modal and notification

DismissEvent(String)

Dismiss an event by id, either a Modal or Notification