Enum stremio_core::runtime::msg::Event

pub enum Event {

    PlayerPlaying {
        context: AnalyticsContext,
        load_time: i64,
    },
    PlayerStopped {
        context: AnalyticsContext,
    },
    PlayerNextVideo {
        context: AnalyticsContext,
        is_binge_enabled: bool,
        is_playing_next_video: bool,
    },
    PlayerEnded {
        context: AnalyticsContext,
        is_binge_enabled: bool,
        is_playing_next_video: bool,
    },
    TraktPlaying {
        context: AnalyticsContext,
    },
    TraktPaused {
        context: AnalyticsContext,
    },
    ProfilePushedToStorage {
        uid: UID,
    },
    LibraryItemsPushedToStorage {
        ids: Vec<String>,
    },
    StreamsPushedToStorage {
        uid: UID,
    },
    SearchHistoryPushedToStorage {
        uid: UID,
    },
    NotificationsPushedToStorage {
        ids: Vec<String>,
    },
    DismissedEventsPushedToStorage {
        uid: UID,
    },
    UserPulledFromAPI {
        uid: UID,
    },
    UserPushedToAPI {
        uid: UID,
    },
    AddonsPulledFromAPI {
        transport_urls: Vec<Url>,
    },
    AddonsPushedToAPI {
        transport_urls: Vec<Url>,
    },
    LibrarySyncWithAPIPlanned {
        uid: UID,
        plan: (Vec<String>, Vec<String>),
    },
    LibraryItemsPushedToAPI {
        ids: Vec<String>,
    },
    LibraryItemsPulledFromAPI {
        ids: Vec<String>,
    },
    UserAuthenticated {
        auth_request: AuthRequest,
    },
    UserAddonsLocked {
        addons_locked: bool,
    },
    UserLibraryMissing {
        library_missing: bool,
    },
    UserLoggedOut {
        uid: UID,
    },
    SessionDeleted {
        auth_key: AuthKey,
    },
    TraktAddonFetched {
        uid: UID,
    },
    TraktLoggedOut {
        uid: UID,
    },
    AddonInstalled {
        transport_url: Url,
        id: String,
    },
    AddonUpgraded {
        transport_url: Url,
        id: String,
    },
    AddonUninstalled {
        transport_url: Url,
        id: String,
    },
    SettingsUpdated {
        settings: Settings,
    },
    LibraryItemAdded {
        id: LibraryItemId,
    },
    LibraryItemRemoved {
        id: LibraryItemId,
    },
    LibraryItemRewinded {
        id: LibraryItemId,
    },
    LibraryItemNotificationsToggled {
        id: LibraryItemId,
    },
    LibraryItemMarkedAsWatched {
        id: LibraryItemId,
        is_watched: bool,
    },
    NotificationsDismissed {
        id: LibraryItemId,
    },
    MagnetParsed {
        magnet: Url,
    },
    TorrentParsed {
        torrent: Vec<u8>,
    },
    PlayingOnDevice {
        device: String,
    },
    Error {
        error: CtxError,
        source: Box<Event>,
    },

}

Those messages are meant to be dispatched by the stremio-core crate and handled by the users of the stremio-core crate and by the stremio-core crate itself.

Variants

PlayerPlaying

Fields

context: AnalyticsContext

load_time: i64

PlayerStopped

Fields

context: AnalyticsContext

PlayerNextVideo

Fields

context: AnalyticsContext

is_binge_enabled: bool

is_playing_next_video: bool

PlayerEnded

Fields

context: AnalyticsContext

is_binge_enabled: bool

is_playing_next_video: bool

TraktPlaying

Fields

context: AnalyticsContext

TraktPaused

Fields

context: AnalyticsContext

ProfilePushedToStorage

Fields

uid: UID

LibraryItemsPushedToStorage

Fields

ids: Vec<String>

StreamsPushedToStorage

Fields

uid: UID

SearchHistoryPushedToStorage

Fields

uid: UID

NotificationsPushedToStorage

Fields

ids: Vec<String>

DismissedEventsPushedToStorage

Fields

uid: UID

UserPulledFromAPI

Fields

uid: UID

UserPushedToAPI

Fields

uid: UID

AddonsPulledFromAPI

Fields

transport_urls: Vec<Url>

AddonsPushedToAPI

Fields

transport_urls: Vec<Url>

LibrarySyncWithAPIPlanned

Fields

uid: UID

plan: (Vec<String>, Vec<String>)

LibraryItemsPushedToAPI

Fields

ids: Vec<String>

LibraryItemsPulledFromAPI

Fields

ids: Vec<String>

UserAuthenticated

Fields

auth_request: AuthRequest

UserAddonsLocked

Fields

addons_locked: bool

UserLibraryMissing

Fields

library_missing: bool

UserLoggedOut

Fields

uid: UID

SessionDeleted

Fields

auth_key: AuthKey

TraktAddonFetched

Fields

uid: UID

TraktLoggedOut

Fields

uid: UID

AddonInstalled

Fields

transport_url: Url

id: String

AddonUpgraded

Fields

transport_url: Url

id: String

AddonUninstalled

Fields

transport_url: Url

id: String

SettingsUpdated

Fields

settings: Settings

LibraryItemAdded

Fields

id: LibraryItemId

LibraryItemRemoved

Fields

id: LibraryItemId

LibraryItemRewinded

Fields

id: LibraryItemId

LibraryItemNotificationsToggled

Fields

id: LibraryItemId

LibraryItemMarkedAsWatched

The LibraryItem with the given id has been marked as watched or unwatched (Overrides the previous watched state)

Fields

id: LibraryItemId

is_watched: bool

NotificationsDismissed

The notifications for the given LibraryItemId have been dismissed

Fields

id: LibraryItemId

MagnetParsed

Fields

magnet: Url

TorrentParsed

Fields

torrent: Vec<u8>

PlayingOnDevice

Fields

device: String

Error

Fields

error: CtxError

source: Box<Event>