Enum stremio_core::runtime::msg::Internal

pub enum Internal {

    CtxAuthResult(AuthRequest, Result<AuthResponse, CtxError>),
    AddonsAPIResult(APIRequest, Result<Vec<Descriptor>, CtxError>),
    UserAPIResult(APIRequest, Result<User, CtxError>),
    LibrarySyncPlanResult(DatastoreRequest, Result<LibraryPlanResponse, CtxError>),
    LibraryPullResult(DatastoreRequest, Result<Vec<LibraryItem>, CtxError>),
    Logout,
    InstallAddon(Descriptor),
    UpdateLibraryItem(LibraryItem),
    ProfileChanged,
    LibraryChanged(bool),
    LinkCodeResult(Result<LinkCodeResponse, LinkError>),
    LinkDataResult(String, Result<LinkDataResponse, LinkError>),
    StreamingServerSettingsResult(Url, Result<StreamingServerSettings, EnvError>),
    StreamingServerBaseURLResult(Url, Result<Url, EnvError>),
    StreamingServerPlaybackDevicesResult(Url, Result<Vec<PlaybackDevice>, EnvError>),
    StreamingServerUpdateSettingsResult(Url, Result<(), EnvError>),
    StreamingServerCreateTorrentResult(String, Result<(), EnvError>),
    StreamingServerPlayOnDeviceResult(String, Result<(), EnvError>),
    ResourceRequestResult(ResourceRequest, Box<Result<ResourceResponse, EnvError>>),
    ManifestRequestResult(Url, Result<Manifest, EnvError>),
    DataExportResult(AuthKey, Result<DataExportResponse, CtxError>),

}

Variants

CtxAuthResult(AuthRequest, Result<AuthResponse, CtxError>)

Result for authenticate to API.

AddonsAPIResult(APIRequest, Result<Vec<Descriptor>, CtxError>)

Result for pull addons from API.

UserAPIResult(APIRequest, Result<User, CtxError>)

Result for pull user from API.

LibrarySyncPlanResult(DatastoreRequest, Result<LibraryPlanResponse, CtxError>)

Result for library sync plan with API.

LibraryPullResult(DatastoreRequest, Result<Vec<LibraryItem>, CtxError>)

Result for pull library items from API.

Logout

Dispatched when expired session is detected

InstallAddon(Descriptor)

Dispatched when addons needs to be installed.

UpdateLibraryItem(LibraryItem)

Dispatched when library item needs to be updated in the memory, storage and API.

ProfileChanged

Dispatched when some of auth, addons or settings changed.

LibraryChanged(bool)

Dispatched when library changes with a flag if its already persisted.

LinkCodeResult(Result<LinkCodeResponse, LinkError>)

Result for loading link code.

LinkDataResult(String, Result<LinkDataResponse, LinkError>)

Result for loading link data.

StreamingServerSettingsResult(Url, Result<StreamingServerSettings, EnvError>)

Result for loading streaming server settings.

StreamingServerBaseURLResult(Url, Result<Url, EnvError>)

Result for loading streaming server base url.

StreamingServerPlaybackDevicesResult(Url, Result<Vec<PlaybackDevice>, EnvError>)

StreamingServerUpdateSettingsResult(Url, Result<(), EnvError>)

Result for updating streaming server settings.

StreamingServerCreateTorrentResult(String, Result<(), EnvError>)

Result for creating a torrent.

StreamingServerPlayOnDeviceResult(String, Result<(), EnvError>)

ResourceRequestResult(ResourceRequest, Box<Result<ResourceResponse, EnvError>>)

Result for fetching resource from addons.

ManifestRequestResult(Url, Result<Manifest, EnvError>)

Result for fetching manifest from addon.

DataExportResult(AuthKey, Result<DataExportResponse, CtxError>)

Result for requesting a dataExport of user data.