Enum stremio_core::types::addon::response::ResourceResponse

pub enum ResourceResponse {
    Metas {
        metas: Vec<MetaItemPreview>,
    },
    MetasDetailed {
        metas_detailed: Vec<MetaItem>,
    },
    Meta {
        meta: MetaItem,
    },
    Streams {
        streams: Vec<Stream>,
    },
    Subtitles {
        subtitles: Vec<Subtitles>,
    },
    Addons {
        addons: Vec<DescriptorPreview>,
    },
}

Resource Response from an addon.

Deserializing the struct from json will skip any invalid Vec items and will skip any unknown to the variants fields.

Variants

Metas

MetaItemPreview response without the videos key of an item

Examples

use stremio_core::types::addon::ResourceResponse;

let null_metas = serde_json::json!({ "metas": null });
let null_metas = serde_json::from_value::<ResourceResponse>(null_metas).expect("Should be valid");

match null_metas {
    ResourceResponse::Metas { metas } => assert!(metas.is_empty()),
    _ => panic!("Whoops, wrong variant!"),
}

Fields

metas: Vec<MetaItemPreview>

MetasDetailed

Fields

metas_detailed: Vec<MetaItem>

Meta

Fields

meta: MetaItem

Streams

Fields

streams: Vec<Stream>

Subtitles

Fields

subtitles: Vec<Subtitles>

Addons

Fields

addons: Vec<DescriptorPreview>