stremio_core::types
Module serde_as_ext

Structs

DefaultOnBool

Deserialize an Option from a bool or the underlying Option. For both true and false values, the Option will be set to None

NumberAsString

SortedVec

UniqueVec

Enums

BoolOrValue 🔒

Traits

SortedVecAdapter

UniqueVecAdapter