Module stremio_core::models::common

Structs

DescriptorLoadable

Fetching addons

ResourceLoadable

When we want to fetch meta items, streams and catalogs

Enums

DescriptorAction

Loadable

ResourceAction

ResourceError

ResourcesAction

Functions

compare_with_priorities

Compare a and b based on provided priority. Priorities should be set based on the type T and i32.

descriptor_update

Request or Load Addon Descriptor on the descriptor argument.

eq_update

resource_update

resource_update_with_vector_content

resources_update

resources_update_with_vector_content