Enum stremio_core::runtime::msg::Internal

pub enum Internal {

    CtxAuthResult(AuthRequest, Result<AuthResponse, CtxError>),
    AddonsAPIResult(APIRequest, Result<Vec<Descriptor>, CtxError>),
    UserAPIResult(APIRequest, Result<User, CtxError>),
    LibrarySyncPlanResult(DatastoreRequest, Result<LibraryPlanResponse, CtxError>),
    LibraryPullResult(DatastoreRequest, Result<Vec<LibraryItem>, CtxError>),
    Logout,
    InstallAddon(Descriptor),
    StreamLoaded {
        stream: Stream,
        meta_id: Option<String>,
        video_id: Option<String>,
        transport_url: Option<Url>,
    },
    UpdateLibraryItem(LibraryItem),
    ProfileChanged,
    LibraryChanged(bool),
    StreamsChanged(bool),
    NotificationsChanged,
    DismissNotificationItem(MetaItemId),
    LinkCodeResult(Result<LinkCodeResponse, LinkError>),
    LinkDataResult(String, Result<LinkDataResponse, LinkError>),
    StreamingServerSettingsResult(Url, Result<Settings, EnvError>),
    StreamingServerBaseURLResult(Url, Result<Url, EnvError>),
    StreamingServerPlaybackDevicesResult(Url, Result<Vec<PlaybackDevice>, EnvError>),
    StreamingServerUpdateSettingsResult(Url, Result<(), EnvError>),
    StreamingServerCreateTorrentResult(String, Result<(), EnvError>),
    StreamingServerPlayOnDeviceResult(String, Result<(), EnvError>),
    StreamingServerStatisticsResult((Url, StatisticsRequest), Result<Statistics, EnvError>),
    ResourceRequestResult(ResourceRequest, Box<Result<ResourceResponse, EnvError>>),
    ManifestRequestResult(Url, Result<Manifest, EnvError>),
    NotificationsRequestResult(ResourceRequest, Box<Result<ResourceResponse, EnvError>>),
    DataExportResult(AuthKey, Result<DataExportResponse, CtxError>),
    LoadLocalSearchResult(Url, Result<Vec<Searchable>, EnvError>),

}

Variants

CtxAuthResult(AuthRequest, Result<AuthResponse, CtxError>)

Result for authenticate to API.

AddonsAPIResult(APIRequest, Result<Vec<Descriptor>, CtxError>)

Result for pull addons from API.

UserAPIResult(APIRequest, Result<User, CtxError>)

Result for pull user from API.

LibrarySyncPlanResult(DatastoreRequest, Result<LibraryPlanResponse, CtxError>)

Result for library sync plan with API.

LibraryPullResult(DatastoreRequest, Result<Vec<LibraryItem>, CtxError>)

Result for pull library items from API.

Logout

Dispatched when expired session is detected

InstallAddon(Descriptor)

Dispatched when addons needs to be installed.

StreamLoaded

Fields

stream: Stream

meta_id: Option<String>

video_id: Option<String>

transport_url: Option<Url>

Dispatched when a new stream is loaded into the Player.

UpdateLibraryItem(LibraryItem)

Dispatched when library item needs to be updated in the memory, storage and API.

ProfileChanged

Dispatched when some of auth, addons or settings changed.

LibraryChanged(bool)

Dispatched when library changes with a flag if its already persisted.

StreamsChanged(bool)

Dispatched when streams bucket changes with a flag if its already persisted.

NotificationsChanged

User notifications have changed

DismissNotificationItem(MetaItemId)

Dismiss all Notifications for a given MetaItemId.

LinkCodeResult(Result<LinkCodeResponse, LinkError>)

Result for loading link code.

LinkDataResult(String, Result<LinkDataResponse, LinkError>)

Result for loading link data.

StreamingServerSettingsResult(Url, Result<Settings, EnvError>)

Result for loading streaming server settings.

StreamingServerBaseURLResult(Url, Result<Url, EnvError>)

Result for loading streaming server base url.

StreamingServerPlaybackDevicesResult(Url, Result<Vec<PlaybackDevice>, EnvError>)

StreamingServerUpdateSettingsResult(Url, Result<(), EnvError>)

Result for updating streaming server settings.

StreamingServerCreateTorrentResult(String, Result<(), EnvError>)

Result for creating a torrent.

StreamingServerPlayOnDeviceResult(String, Result<(), EnvError>)

StreamingServerStatisticsResult((Url, StatisticsRequest), Result<Statistics, EnvError>)

ResourceRequestResult(ResourceRequest, Box<Result<ResourceResponse, EnvError>>)

Result for fetching resource from addons.

ManifestRequestResult(Url, Result<Manifest, EnvError>)

Result for fetching manifest from addon.

NotificationsRequestResult(ResourceRequest, Box<Result<ResourceResponse, EnvError>>)

TODO: write some obvious comment about what it is

DataExportResult(AuthKey, Result<DataExportResponse, CtxError>)

Result for requesting a dataExport of user data.

LoadLocalSearchResult(Url, Result<Vec<Searchable>, EnvError>)

The result of querying the data for LocalSearch