Enum stremio_core::runtime::msg::Internal

pub enum Internal {

    CtxAuthResult(AuthRequest, Result<CtxAuthResponse, CtxError>),
    AddonsAPIResult(APIRequest, Result<Vec<Descriptor>, CtxError>),
    UserAPIResult(APIRequest, Result<User, CtxError>),
    LibrarySyncPlanResult(DatastoreRequest, Result<LibraryPlanResponse, CtxError>),
    LibraryPullResult(DatastoreRequest, Result<Vec<LibraryItem>, CtxError>),
    Logout,
    InstallTraktAddon,
    InstallAddon(Descriptor),
    UninstallAddon(Descriptor),
    UninstallTraktAddon,
    StreamLoaded {
        stream: Stream,
        stream_request: Option<ResourceRequest>,
        meta_item: ResourceLoadable<MetaItem>,
    },
    StreamStateChanged {
        state: StreamItemState,
        stream_request: Option<ResourceRequest>,
        meta_request: Option<ResourceRequest>,
    },
    CatalogsWithExtraSearch {
        query: String,
    },
    UpdateLibraryItem(LibraryItem),
    ProfileChanged,
    LibraryChanged(bool),
    StreamsChanged(bool),
    SearchHistoryChanged,
    NotificationsChanged,
    PullNotifications,
    DismissNotificationItem(LibraryItemId),
    LinkCodeResult(Result<LinkCodeResponse, LinkError>),
    LinkDataResult(String, Result<LinkDataResponse, LinkError>),
    StreamingServerSettingsResult(Url, Result<SettingsResponse, EnvError>),
    StreamingServerBaseURLResult(Url, Result<Url, EnvError>),
    StreamingServerPlaybackDevicesResult(Url, Result<Vec<PlaybackDevice>, EnvError>),
    StreamingServerNetworkInfoResult(Url, Result<NetworkInfo, EnvError>),
    StreamingServerDeviceInfoResult(Url, Result<DeviceInfo, EnvError>),
    StreamingServerUpdateSettingsResult(Url, Result<(), EnvError>),
    StreamingServerCreateTorrentResult(String, Result<(), EnvError>),
    StreamingServerPlayOnDeviceResult(String, Result<(), EnvError>),
    StreamingServerGetHTTPSResult(Url, Result<GetHTTPSResponse, EnvError>),
    StreamingServerStatisticsResult((Url, StatisticsRequest), Result<Option<Statistics>, EnvError>),
    ResourceRequestResult(ResourceRequest, Box<Result<ResourceResponse, EnvError>>),
    ManifestRequestResult(Url, Result<Manifest, EnvError>),
    NotificationsRequestResult(ResourceRequest, Box<Result<ResourceResponse, EnvError>>),
    DataExportResult(AuthKey, Result<DataExportResponse, CtxError>),
    SeekLogsResult(SeekLogRequest, Result<SuccessResponse, CtxError>),
    SkipGapsResult(SkipGapsRequest, Result<SkipGapsResponse, CtxError>),
    LoadLocalSearchResult(Url, Result<Vec<Searchable>, EnvError>),
    GetModalResult(APIRequest, Result<Option<GetModalResponse>, CtxError>),
    GetNotificationResult(APIRequest, Result<Option<GetNotificationResponse>, CtxError>),
    DismissedEventsChanged,

}

Variants

CtxAuthResult(AuthRequest, Result<CtxAuthResponse, CtxError>)

Result for authenticate to API.

AddonsAPIResult(APIRequest, Result<Vec<Descriptor>, CtxError>)

Result for pull addons from API.

UserAPIResult(APIRequest, Result<User, CtxError>)

Result for pull user from API.

LibrarySyncPlanResult(DatastoreRequest, Result<LibraryPlanResponse, CtxError>)

Result for library sync plan with API.

LibraryPullResult(DatastoreRequest, Result<Vec<LibraryItem>, CtxError>)

Result for pull library items from API.

Logout

Dispatched when expired session is detected

InstallTraktAddon

Internal event dispatched on user action or login to install the addon if it’s not present

InstallAddon(Descriptor)

Dispatched when addons needs to be installed.

UninstallAddon(Descriptor)

Dispatched when addons needs to be uninstalled.

UninstallTraktAddon

StreamLoaded

Dispatched when a new stream is loaded into the Player.

Fields

stream: Stream

stream_request: Option<ResourceRequest>

meta_item: ResourceLoadable<MetaItem>

StreamStateChanged

Dispatched when stream item’s state has changed

Fields

state: StreamItemState

stream_request: Option<ResourceRequest>

meta_request: Option<ResourceRequest>

CatalogsWithExtraSearch

Dispatched when requesting search on catalogs.

Fields

query: String

UpdateLibraryItem(LibraryItem)

Dispatched when library item needs to be updated in the memory, storage and API.

ProfileChanged

Dispatched when some of auth, addons or settings changed.

LibraryChanged(bool)

Dispatched when library changes with a flag if its already persisted.

StreamsChanged(bool)

Dispatched when streams bucket changes with a flag if its already persisted.

SearchHistoryChanged

Search history has changed.

NotificationsChanged

User notifications have changed

PullNotifications

Pulling of notifications triggered either by the user (with an action) or internally in core.

DismissNotificationItem(LibraryItemId)

Dismiss all Notifications for a given MetaItemId.

LinkCodeResult(Result<LinkCodeResponse, LinkError>)

Result for loading link code.

LinkDataResult(String, Result<LinkDataResponse, LinkError>)

Result for loading link data.

StreamingServerSettingsResult(Url, Result<SettingsResponse, EnvError>)

Result for loading streaming server settings.

StreamingServerBaseURLResult(Url, Result<Url, EnvError>)

Result for loading streaming server base url.

StreamingServerPlaybackDevicesResult(Url, Result<Vec<PlaybackDevice>, EnvError>)

StreamingServerNetworkInfoResult(Url, Result<NetworkInfo, EnvError>)

StreamingServerDeviceInfoResult(Url, Result<DeviceInfo, EnvError>)

StreamingServerUpdateSettingsResult(Url, Result<(), EnvError>)

Result for updating streaming server settings.

StreamingServerCreateTorrentResult(String, Result<(), EnvError>)

Result for creating a torrent.

StreamingServerPlayOnDeviceResult(String, Result<(), EnvError>)

Result for playing on device.

StreamingServerGetHTTPSResult(Url, Result<GetHTTPSResponse, EnvError>)

StreamingServerStatisticsResult((Url, StatisticsRequest), Result<Option<Statistics>, EnvError>)

Result for streaming server statistics.

Server will return None (or null) in response for Statistics`, when stream has been fully loaded up to 100%

ResourceRequestResult(ResourceRequest, Box<Result<ResourceResponse, EnvError>>)

Result for fetching resource from addons.

ManifestRequestResult(Url, Result<Manifest, EnvError>)

Result for fetching manifest from addon.

NotificationsRequestResult(ResourceRequest, Box<Result<ResourceResponse, EnvError>>)

TODO: write some obvious comment about what it is

DataExportResult(AuthKey, Result<DataExportResponse, CtxError>)

Result for requesting a dataExport of user data.

SeekLogsResult(SeekLogRequest, Result<SuccessResponse, CtxError>)

Result for submitting SeekLogs request for a played stream.

Applicable only to movie series and torrents.

SkipGapsResult(SkipGapsRequest, Result<SkipGapsResponse, CtxError>)

Retrieve the skip gaps for skipping intro and outro.

LoadLocalSearchResult(Url, Result<Vec<Searchable>, EnvError>)

The result of querying the data for LocalSearch

GetModalResult(APIRequest, Result<Option<GetModalResponse>, CtxError>)

Result for getModal request

GetNotificationResult(APIRequest, Result<Option<GetNotificationResponse>, CtxError>)

Result for getNotification request

DismissedEventsChanged

When dismissed events changed