stremio_core
Module types

Modules

addon

api

empty_string_as_null

events

library

notifications

player

profile

query_params_encode 🔒

resource

search_history

serde_as_ext 🔒

serde_ext 🔒

streaming_server

streams

torrent

true 🔒

Structs

DefaultOnBool

Deserialize an Option from a bool or the underlying Option. For both true and false values, the Option will be set to None

NumberAsString

SortedVec

True

UniqueVec

Traits

SortedVecAdapter

UniqueVecAdapter

Functions

query_params_encode