stremio_core::types
Module addon

Modules

descriptor 🔒

manifest 🔒

request 🔒

response 🔒

Structs

Descriptor

Addon descriptor

DescriptorFlags

DescriptorPreview

ExtraProp

ExtraValue

Manifest

ManifestBehaviorHints

ManifestCatalog

ManifestPreview

OptionsLimit

ResourcePath

The full resource path, query, etc. for Addon requests

ResourceRequest

ResourceResponseCache

Resource response that handles the optional cache values defined in the SDK. This is a shim struct to avoid the custom Deserialize impl that skips those fields for the ResourceResponse enum.

Version

Re-export the semver::Version SemVer version as defined by https://semver.org.

Enums

AggrRequest

ExtraType

ManifestExtra

ManifestResource

Resources supported by the addon.

ResourceResponse

Resource Response from an addon.

Traits

ExtraExt