stremio_core::types
Module resource

Modules

meta_item 🔒

stream 🔒

subtitles 🔒

Structs

Link

MetaItem

MetaItemBehaviorHints

MetaItemPreview

SeriesInfo

Stream

Examples

StreamBehaviorHints

See https://github.com/Stremio/stremio-addon-sdk/blob/master/docs/api/responses/stream.md#additional-properties-to-provide-information--behaviour-flags for documentation

StreamProxyHeaders

Subtitles

Video

Enums

PosterShape

StreamSource

Examples

Type Aliases

MetaItemId

The MetaItem Id type to improve the readability of the code.

VideoId

The Video Id type to improve the readability of the code.