Module stremio_core::types::resource

Modules

meta_item 🔒

stream 🔒

subtitles 🔒

Structs

Link

MetaItem

MetaItemBehaviorHints

MetaItemPreview

SeriesInfo

Stream

StreamBehaviorHints

StreamProxyHeaders

Subtitles

Video

Enums

PosterShape

StreamSource

Type Aliases

MetaItemId

The MetaItem Id type to improve the readability of the code.

VideoId

The Video Id type to improve the readability of the code.